Unified materialized binary
