+++ bin/bump-version
@@ -38,7 +38,7 @@ version=${1#v}
 
 sed -i.bak \
     "s/^version = .*/version = \"$version\"/" \
-    src/{clusterd,environmentd,persist-client,testdrive,catalog-debug,balancerd,orchestratord}/Cargo.toml
+    src/{clusterd,environmentd,materialized,persist-client,testdrive,catalog-debug,balancerd,orchestratord}/Cargo.toml
 
 if ! [[ "$version" = *-dev* ]]; then
     sed -i.bak \
@@ -58,7 +58,7 @@ else
     done
 fi
 
-rm -f src/{clusterd,environmentd,persist-client,testdrive,catalog-debug,balancerd,orchestratord}/Cargo.toml.bak LICENSE.bak
+rm -f src/{clusterd,environmentd,materialized,persist-client,testdrive,catalog-debug,balancerd,orchestratord}/Cargo.toml.bak LICENSE.bak
 
 cargo update --workspace
