+++ ci/test/lint-deps/aarch64-unknown-linux-gnu-default
@@ -1,17 +1,17 @@
 # generated by ci/test/lint-deps.sh -- see ci/test/lint-deps/README.md for details
 tikv_jemalloc_ctl
-0::default,mz-alloc-default,tokio-console
+0:mz_clusterd:default,mz-alloc-default,tokio-console
 1:mz_alloc:default,jemalloc,mz-prof,tikv-jemallocator,workspace-hack
 2:mz_prof:jemalloc,jemalloc_pprof,tikv-jemalloc-ctl
 3:jemalloc_pprof:
 4:tikv_jemalloc_ctl:default,stats,use_std
 3:tikv_jemalloc_ctl:default,stats,use_std (*)
 tikv_jemallocator
-0::default,mz-alloc-default,tokio-console
+0:mz_clusterd:default,mz-alloc-default,tokio-console
 1:mz_alloc:default,jemalloc,mz-prof,tikv-jemallocator,workspace-hack
 2:tikv_jemallocator:background_threads,background_threads_runtime_support,default,profiling,stats,unprefixed_malloc_on_supported_platforms
 tikv_jemalloc_sys
-0::default,mz-alloc-default,tokio-console
+0:mz_clusterd:default,mz-alloc-default,tokio-console
 1:mz_alloc:default,jemalloc,mz-prof,tikv-jemallocator,workspace-hack
 2:mz_prof:jemalloc,jemalloc_pprof,tikv-jemalloc-ctl
 3:jemalloc_pprof:
