+++ src/environmentd/Cargo.toml
@@ -189,7 +189,7 @@ normal = ["mz-alloc-default", "sysctl", "workspace-hack"]
 compile_data = ["src/http/static/**/*", "templates/*"]
 
 [package.metadata.cargo-gazelle.build]
-data = ["src/bin/environmentd/sys.c"]
+data = ["src/environmentd/sys.c"]
 
 [package.metadata.cargo-gazelle.test.server]
 env = { INSTA_WORKSPACE_ROOT = "." }
