+++ src/environmentd/src/bin/environmentd.rs
@@ -7,14 +7,6 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0.
 
-use std::env;
-
-fn main() -> Result<(), anyhow::Error> {
-    println!("cargo:rustc-env=TARGET_TRIPLE={}", env::var("TARGET")?);
-
-    cc::Build::new()
-        .file("src/environmentd/sys.c")
-        .compile("environmentd_sys");
-
-    mz_npm::ensure()
+fn main() {
+    mz_environmentd::environmentd::main();
 }
