+++ src/materialized/BUILD.bazel
@@ -1,0 +1,37 @@
+# Code generated by cargo-gazelle DO NOT EDIT
+
+# Copyright Materialize, Inc. and contributors. All rights reserved.
+#
+# Use of this software is governed by the Business Source License
+# included in the LICENSE file at the root of this repository.
+#
+# As of the Change Date specified in that file, in accordance with
+# the Business Source License, use of this software will be governed
+# by the Apache License, Version 2.0.
+
+load("@crates_io//:defs.bzl", "aliases", "all_crate_deps")
+load("@rules_rust//rust:defs.bzl", "rust_binary")
+
+package(default_visibility = ["//visibility:public"])
+
+rust_binary(
+    name = "materialized",
+    srcs = glob(["src/**/*.rs"]),
+    aliases = aliases(
+        normal = True,
+        proc_macro = True,
+    ),
+    compile_data = [],
+    crate_root = "src/bin/materialized.rs",
+    data = [],
+    env = {},
+    features = [],
+    proc_macro_deps = [] + all_crate_deps(proc_macro = True),
+    rustc_env = {},
+    rustc_flags = [],
+    version = "0.126.0-dev.0",
+    deps = [
+        "//src/clusterd:mz_clusterd",
+        "//src/environmentd:mz_environmentd",
+    ] + all_crate_deps(normal = True),
+)
