+++ src/materialized/ci/.gitignore
@@ -1,2 +1,3 @@
 /clusterd
 /environmentd
+/materialized
