+++ src/materialized/ci/entrypoint.sh
