+++ src/materialized/src/bin/materialized.rs
@@ -7,14 +7,13 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0.
 
-use std::env;
+use std::ffi::OsStr;
 
-fn main() -> Result<(), anyhow::Error> {
-    println!("cargo:rustc-env=TARGET_TRIPLE={}", env::var("TARGET")?);
-
-    cc::Build::new()
-        .file("src/environmentd/sys.c")
-        .compile("environmentd_sys");
-
-    mz_npm::ensure()
+fn main() {
+    let binary = std::env::current_exe().expect("known executable");
+    match binary.file_name().and_then(OsStr::to_str) {
+        Some("clusterd") => mz_clusterd::main(),
+        Some("materialized") => mz_environmentd::environmentd::main(),
+        other => panic!("Unknown executable: {other:?}"),
+    }
 }
